Move into a React-ish project structure
