+++ src/SelectedVersionDetails/LicensingPage/LicensingDisplay/LicensingDisplay.tsx
