+++ src/SelectedVersionDetails/LicensingPage/LicensingPage.tsx
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 import * as React from 'react';
-import LicensingDisplay from './LicensingDisplay';
+import LicensingDisplay from './LicensingDisplay/LicensingDisplay';
 import Table from 'react-bootstrap/Table';
-import { VersionsContextConsumer } from './context/versions-context';
+import { VersionsContextConsumer } from '../../context/versions-context';
 
 type Props = {
 }
